+++ crates/holochain/src/core/workflow/sys_validation_workflow.rs
@@ -91,6 +91,7 @@ use holochain_keystore::MetaLairClient;
 use holochain_p2p::GenericNetwork;
 use holochain_p2p::HolochainP2pDnaT;
 use holochain_sqlite::prelude::*;
+use holochain_sqlite::sql::sql_cell::ACTION_HASH_BY_PREV;
 use holochain_state::prelude::*;
 use parking_lot::Mutex;
 use rusqlite::Transaction;
@@ -124,6 +125,7 @@ pub async fn sys_validation_workflow<Network: HolochainP2pDnaT + 'static>(
     workspace: Arc<SysValidationWorkspace>,
     current_validation_dependencies: Arc<Mutex<ValidationDependencies>>,
     trigger_app_validation: TriggerSender,
+    trigger_publish: TriggerSender,
     trigger_self: TriggerSender,
     network: Network,
     config: Arc<ConductorConfig>,
@@ -145,6 +147,15 @@ pub async fn sys_validation_workflow<Network: HolochainP2pDnaT + 'static>(
         tracing::debug!("Sys validation accepted {} ops", outcome_summary.accepted);
 
         trigger_app_validation.trigger(&"sys_validation_workflow");
+    }
+
+    if outcome_summary.warranted > 0 {
+        tracing::debug!(
+            "Sys validation created {} warrants",
+            outcome_summary.warranted
+        );
+
+        trigger_publish.trigger(&"sys_validation_workflow");
     }
 
     // Now go to the network to try to fetch missing dependencies
@@ -275,14 +286,31 @@ async fn sys_validation_workflow_inner(
         }
     }
 
-    let (summary, invalid_ops) = workspace
+    let mut warrants = vec![];
+
+    let (mut summary, invalid_ops, forked_pairs) = workspace
         .dht_db
         .write_async(move |txn| {
             let mut summary = OutcomeSummary::default();
             let mut invalid_ops = vec![];
+            let mut forked_pairs = vec![];
 
             for (hashed_op, outcome) in validation_outcomes {
                 let (op, op_hash) = hashed_op.into_inner();
+                let op_type = op.get_type();
+
+                if let DhtOp::ChainOp(chain_op) = &op {
+                    // Author a ChainFork warrant if fork is detected
+                    let action = chain_op.action();
+                    if let Some(forked_action) = detect_fork(txn, &action)? {
+                        let signature = chain_op.signature().clone();
+                        let action_hash = action.to_hash();
+                        forked_pairs.push((
+                            action.author().clone(),
+                            ((action_hash, signature), forked_action),
+                        ));
+                    }
+                }
 
                 // This is an optimization to skip app validation and integration for ops that are
                 // rejected and don't have dependencies.
@@ -291,7 +319,16 @@ async fn sys_validation_workflow_inner(
                 match outcome {
                     Outcome::Accepted => {
                         summary.accepted += 1;
-                        put_validation_limbo(txn, &op_hash, ValidationStage::SysValidated)?;
+                        match op_type {
+                            DhtOpType::Chain(_) => {
+                                put_validation_limbo(txn, &op_hash, ValidationStage::SysValidated)?
+                            }
+                            DhtOpType::Warrant(_) => {
+                                // XXX: integrate accepted warrants immediately, because we don't
+                                //      want them to go to app validation.
+                                put_integrated(txn, &op_hash, ValidationStatus::Valid)?
+                            }
+                        };
                         aitia::trace!(&hc_sleuth::Event::SysValidated {
                             by: sleuth_id.clone(),
                             op: op_hash
@@ -314,15 +351,13 @@ async fn sys_validation_workflow_inner(
                     }
                 }
             }
-            WorkflowResult::Ok((summary, invalid_ops))
+            WorkflowResult::Ok((summary, invalid_ops, forked_pairs))
         })
         .await?;
 
-    let mut warrants = vec![];
-
     for (_, op) in invalid_ops {
         if let Some(chain_op) = op.as_chain_op() {
-            let warrant_op = crate::core::workflow::sys_validation_workflow::make_warrant_op_inner(
+            let warrant_op = crate::core::workflow::sys_validation_workflow::make_invalid_chain_warrant_op_inner(
                 &keystore,
                 representative_agent.clone(),
                 chain_op,
@@ -333,11 +368,19 @@ async fn sys_validation_workflow_inner(
         }
     }
 
+    for (author, pair) in forked_pairs {
+        let warrant_op =
+            make_fork_warrant_op_inner(&keystore, representative_agent.clone(), author, pair)
+                .await?;
+        warrants.push(warrant_op);
+    }
+
     workspace
         .authored_db
         .write_async(move |txn| {
             for warrant_op in warrants {
                 insert_op(txn, &warrant_op)?;
+                summary.warranted += 1;
             }
             StateMutationResult::Ok(())
         })
@@ -505,8 +548,8 @@ fn get_dependency_hashes_from_ops(ops: impl Iterator<Item = DhtOpHashed>) -> Vec
                     }
                     _ => None,
                 },
-                DhtOp::WarrantOp(op) => match &op.warrant {
-                    Warrant::ChainIntegrity(warrant) => match warrant {
+                DhtOp::WarrantOp(op) => match &op.proof {
+                    WarrantProof::ChainIntegrity(warrant) => match warrant {
                         ChainIntegrityWarrant::InvalidChainOp {
                             action: (action_hash, _),
                             ..
@@ -732,8 +775,8 @@ async fn validate_warrant_op(
     _dna_def: &DnaDefHashed,
     validation_dependencies: Arc<Mutex<ValidationDependencies>>,
 ) -> SysValidationResult<()> {
-    match &op.warrant {
-        Warrant::ChainIntegrity(warrant) => match warrant {
+    match &op.proof {
+        WarrantProof::ChainIntegrity(warrant) => match warrant {
             ChainIntegrityWarrant::InvalidChainOp {
                 action: (action_hash, action_sig),
                 action_author,
@@ -749,8 +792,8 @@ async fn validate_warrant_op(
                         })?;
 
                     if action.author() != action_author {
-                        return Err(ValidationOutcome::InvalidWarrantOp(
-                            op.clone(),
+                        return Err(ValidationOutcome::InvalidWarrant(
+                            op.warrant().clone(),
                             "action author mismatch".into(),
                         )
                         .into());
@@ -779,8 +822,8 @@ async fn validate_warrant_op(
 
                     // chain_author basis must match the author of the action
                     if action1.author() != chain_author {
-                        return Err(ValidationOutcome::InvalidWarrantOp(
-                            op.clone(),
+                        return Err(ValidationOutcome::InvalidWarrant(
+                            op.warrant().clone(),
                             "chain author mismatch".into(),
                         )
                         .into());
@@ -788,8 +831,8 @@ async fn validate_warrant_op(
 
                     // Both actions must be by same author
                     if action1.author() != action2.author() {
-                        return Err(ValidationOutcome::InvalidWarrantOp(
-                            op.clone(),
+                        return Err(ValidationOutcome::InvalidWarrant(
+                            op.warrant().clone(),
                             "action pair author mismatch".into(),
                         )
                         .into());
@@ -802,8 +845,8 @@ async fn validate_warrant_op(
                     // Using seq numbers makes it easier to detect and prove a fork, but using prev_action
                     // prevents the attack.
                     if action1.prev_action() != action2.prev_action() {
-                        return Err(ValidationOutcome::InvalidWarrantOp(
-                            op.clone(),
+                        return Err(ValidationOutcome::InvalidWarrant(
+                            op.warrant().clone(),
                             "action pair seq mismatch".into(),
                         )
                         .into());
@@ -1321,7 +1364,7 @@ pub async fn make_warrant_op(
 ) -> WorkflowResult<DhtOpHashed> {
     let keystore = conductor.keystore();
     let warrant_author = get_representative_agent(conductor, dna_hash).expect("TODO: handle");
-    make_warrant_op_inner(keystore, warrant_author, op, validation_type).await
+    make_invalid_chain_warrant_op_inner(keystore, warrant_author, op, validation_type).await
 }
 
 /// Gets an arbitrary agent with a cell running the given DNA, needed for processes
@@ -1335,7 +1378,7 @@ pub fn get_representative_agent(conductor: &Conductor, dna_hash: &DnaHash) -> Op
         .map(|id| id.agent_pubkey().clone())
 }
 
-pub async fn make_warrant_op_inner(
+pub async fn make_invalid_chain_warrant_op_inner(
     keystore: &MetaLairClient,
     warrant_author: AgentPubKey,
     op: &ChainOp,
@@ -1345,12 +1388,13 @@ pub async fn make_warrant_op_inner(
     let action_author = action.author().clone();
     tracing::warn!("Authoring warrant for invalid op authored by {action_author}");
 
-    let warrant = Warrant::ChainIntegrity(ChainIntegrityWarrant::InvalidChainOp {
+    let proof = WarrantProof::ChainIntegrity(ChainIntegrityWarrant::InvalidChainOp {
         action_author,
         action: (action.to_hash().clone(), op.signature().clone()),
         validation_type,
     });
-    let warrant_op = WarrantOp::author(keystore, warrant_author, warrant)
+    let warrant = Warrant::new(proof, warrant_author, Timestamp::now());
+    let warrant_op = WarrantOp::sign(keystore, warrant)
         .await
         .map_err(|e| super::WorkflowError::Other(e.into()))?;
     let op: DhtOp = warrant_op.into();
@@ -1358,11 +1402,65 @@ pub async fn make_warrant_op_inner(
     Ok(op)
 }
 
+pub async fn make_fork_warrant_op_inner(
+    keystore: &MetaLairClient,
+    warrant_author: AgentPubKey,
+    chain_author: AgentPubKey,
+    action_pair: ((ActionHash, Signature), (ActionHash, Signature)),
+) -> WorkflowResult<DhtOpHashed> {
+    debug_assert_ne!(action_pair.0 .0, action_pair.1 .0);
+    tracing::warn!(
+        "Authoring warrant for chain fork by {chain_author}. Action hashes: ({}, {})",
+        action_pair.0 .0,
+        action_pair.1 .0
+    );
+
+    let warrant = Warrant::new(
+        WarrantProof::ChainIntegrity(ChainIntegrityWarrant::ChainFork {
+            chain_author,
+            action_pair,
+        }),
+        warrant_author,
+        Timestamp::now(),
+    );
+    let warrant_op = WarrantOp::sign(keystore, warrant)
+        .await
+        .map_err(|e| super::WorkflowError::Other(e.into()))?;
+    let op: DhtOp = warrant_op.into();
+    let op = op.into_hashed();
+    Ok(op)
+}
+
+pub fn detect_fork(
+    txn: &mut Transaction<'_>,
+    action: &Action,
+) -> StateQueryResult<Option<(ActionHash, Signature)>> {
+    let mut statement = txn.prepare(ACTION_HASH_BY_PREV)?;
+    statement
+        .query_row(
+            named_params! {
+                ":prev_hash": action.prev_action(),
+                ":hash": action.to_hash(),
+            },
+            |row| {
+                let hash: ActionHash = row.get("hash")?;
+                let action_blob: Vec<u8> = row.get("blob")?;
+                Ok((hash, action_blob))
+            },
+        )
+        .optional()?
+        .map(|(hash, action_blob)| {
+            from_blob::<SignedAction>(action_blob).map(|action| (hash, action.signature().clone()))
+        })
+        .transpose()
+}
+
 #[derive(Debug, Clone)]
 struct OutcomeSummary {
     accepted: usize,
     missing: usize,
     rejected: usize,
+    warranted: usize,
 }
 
 impl OutcomeSummary {
@@ -1371,6 +1469,7 @@ impl OutcomeSummary {
             accepted: 0,
             missing: 0,
             rejected: 0,
+            warranted: 0,
         }
     }
 }